// frontend/src/enums/ProblemTags.tsx
export enum ProblemTags {
    IMPLEMENTATION = "IMPLEMENTATION",
    MATH = "MATH",
    GREEDY = "GREEDY",
    TWO_POINTERS = "TWO_POINTERS",
    STRINGS = "STRINGS",
    SORTING = "SORTING",
    DATA_STRUCTURES = "DATA_STRUCTURES",
    GRAPH_THEORY = "GRAPH_THEORY",
    DP = "DP",
    BRUTE_FORCE = "BRUTE_FORCE",
    BINARY_SEARCH = "BINARY_SEARCH",
    TREES = "TREES",
    DFS_AND_SIMILAR = "DFS_AND_SIMILAR",
    BFS = "BFS",
    COMBINATORICS = "COMBINATORICS",
    GEOMETRY = "GEOMETRY",
    HASHING = "HASHING",
    DSU = "DSU",
    HEAPS = "HEAPS",
  }
  